+++ qgitclonerepositorydialog.h
@@ -22,7 +22,7 @@ private slots:
     void on_pushButton_close_clicked();
     void repositoryCloneReply(QDir path, int error);
     void repositoryCloneTransferReply(unsigned int total_objects, unsigned int indexed_objects, unsigned int received_objects, unsigned int local_objects, unsigned int total_deltas, unsigned int indexed_deltas, size_t received_bytes);
-    void repositoryCloneProgressReply(QDir path, int completed_steps, int total);
+    void repositoryCloneProgressReply(QDir path, size_t completed_steps, size_t total);
 
 private:
     Ui::QGitCloneRepositoryDialog *ui;
